+++ src/webview/lib/RecipeWebview.js
@@ -4,12 +4,8 @@ import { pathExistsSync, readFile } from 'fs-extra';
 const debug = require('debug')('Ferdi:Plugin:RecipeWebview');
 
 class RecipeWebview {
-  constructor(notificationsHandler) {
-    this.countCache = {
-      direct: 0,
-      indirect: 0,
-    };
-
+  constructor(badgeHandler, notificationsHandler) {
+    this.badgeHandler = badgeHandler;
     this.notificationsHandler = notificationsHandler;
 
     ipcRenderer.on('poll', () => {
@@ -47,24 +43,7 @@ class RecipeWebview {
    *                          me directly to me eg. in a channel
    */
   setBadge(direct = 0, indirect = 0) {
-    if (this.countCache.direct === direct
-      && this.countCache.indirect === indirect) return;
-
-    // Parse number to integer
-    // This will correct errors that recipes may introduce, e.g.
-    // by sending a String instead of an integer
-    const directInt = parseInt(direct, 10);
-    const indirectInt = parseInt(indirect, 10);
-
-    const count = {
-      direct: Math.max(directInt, 0),
-      indirect: Math.max(indirectInt, 0),
-    };
-
-    ipcRenderer.sendToHost('message-counts', count);
-    Object.assign(this.countCache, count);
-
-    debug('Sending badge count to host', count);
+    this.badgeHandler.setBadge(direct, indirect);
   }
 
   /**
